strip ansi escape sequences from compiler error
