+++ lib/mix/tasks/tableau.build.ex
@@ -51,7 +51,7 @@ defmodule Mix.Tasks.Tableau.Build do
           Map.merge(page, %{body: content, permalink: permalink})
         rescue
           exception ->
-            reraise Tableau.BuildException, [page: page, exception: exception], __STACKTRACE__
+            reraise TableauDevServer.BuildException, [page: page, exception: exception], __STACKTRACE__
         end
       end)
       |> Stream.map(fn
